refactor(rib): extract read-only query handlers
